% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{getCellTypes}
\alias{getCellTypes}
\title{Obten el tipo de una celula}
\usage{
getCellTypes(x)
}
\arguments{
\item{x}{Nombre de una célula.}
}
\value{
Devuelve el tipo de célula.
}
\description{
La función obtiene el tipo de célula a partir de un valor de entrada.
}
\section{Warning}{

En esta versión el tipo de célula es siempre una neurona.
}

\examples{
getCellTypes(motoneurona)
}
